refactor: move chat/component to /component
