+++ components/chat/inputMessage.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import Image from 'next/image';
-import send from '../../../public/Chat/send.png';
+import send from '../../public/Chat/send.png';
 
 const InputContainer: React.FC<{
   messageText: string;
