+++ components/chat/lobbyHeader.tsx
@@ -3,9 +3,9 @@ import Image from 'next/image';
 import React, { useState, useEffect, useRef } from 'react';
 import { useRouter } from 'next/router';
 import CreateRoomModal from './createRoomModal';
-import info from '../../../public/Icon/info.png';
-import plus from '../../../public/Icon/plus.png';
-import home from '../../../public/Icon/home.png';
+import info from '../../public/Icon/info.png';
+import plus from '../../public/Icon/plus.png';
+import home from '../../public/Icon/home.png';
 
 const Header = () => {
   const [isOpenModal, setOpenModal] = useState<boolean>(false);
